Changed Sim function to Sim class. Started to move variables and methods from LLG to the new Simulation class. Created a driver class so every driver, e.g. LLG, STT LLG, etc., inherit from this driver, focusing only on solving the dynamical equation. The Simulation class will keep methods related to the system properties, such as average magnetisation, skyrmion number calculation, etc. We still have a debate about the field arrays (spin, field, alpha, etc), if they should be kept on Sim or move them to the Driver class or maybe we can just reference the arrays from the driver classes (we still need research on this). The same applies to functions that could be useful in the Sim class, such as SaveVTK
